plugins/rag-plugin/requirements.txt
# Runtime
openai
chromadb
pypdf
python-docx
Office365-REST-Python-Client

# Dev
pytest
pytest-cov
